WT-5319 Avoid clearing the saved last-key when no instantiated key (#5041)
Reconciliation has some fast-path code to work on prefix-compressed
to avoid figuring out each key from scratch which would require
going back some number of K/V pairs to key that's not
prefix-compressed and then rolling forward), we use the complete
key we just built for the last K/V pair's reconciliation, and
apply the current key's prefix-compression change to it, saving
us quite a bit of work. The change in WT-4642 clears that saved,
last-key information, and it leads to a performance drop compared
to 4.0 version. More details are available in SERVER-44991.