Check WT_TXN_ERROR before proceeding with any call: This isn't a correctness check, if the WT_TXN_ERROR flag is set, we'll rollback the transaction no matter what the application does. The reason for the change is to ensure we log an error so the application knows it made a mistake, and second, to do it as soon as possible so the application knows where to look. Assert in diagnostic builds that's the case.     That fix requires a change to complex table cursor APPLY_CG macro so it immediately fails if it sees any error other than WT_NOTFOUND, so rather than repeatedly working through the list of table cursors, it quits as soon as there's anything other than a not-found error.
Fix a problem where attempting an operation after a transaction has been prepared
can make it impossible to commit the prepared transaction. In short, attempting
an unsupported operation after a transaction is successfully prepared cannot set
the WT_TXN_ERROR flag, otherwise a subsequent attempt to commit the transaction
will instead roll it back. This requires we check if a transaction is prepared
(and return an error if it's not), before using the standard XXX_CONF macros,
because the corresponding API_END macro will will set the WT_TXN_ERROR flag if
the ret variable is non-zero.
    That fix requires a bunch of cursor function restructuring so functions
that return void can ignore the returned value from the prepared transaction
check function. In general, the pattern is to have a worker function that returns
an int and does all of the work, and a function that calls the worker function
and ignores its return. We need that pattern for other reasons in some places,
so this change just makes that a more common pattern to follow.
    That fix also requires passing in the operation "name" to the checking
function because we won't have the WT_SESSION handle set up before the call.
    That fix also requires minor changes to the static test suite where we were
attempting operations after a transactional call had already failed.

Move __wt_txn_context_prepare_check() and __wt_txn_context_check() from misc.i
to txn.i, the transaction subsystem has its own inlined function file.