WT-5300 checkpoint obsoleting internal page
Whenever checkpoint is flushing the internal page, it scans the
entire page to find out the obsolete child pages and mark them
as obsolete, during the reconciliation of the internal page, all
the child pages are removed. If the internal page is a clean page
and empty, it will removed like a leaf page, otherwise, this page
is processed like a dirty page to remove the child pages first
and later the internal page is removed in the next checkpoint.

Apart of the above, the tree walk reads the internal pages into
memory if they don't exist in memory and they are from history
store for the obsolete page removal.