WT-5324 review of error handling in the WiredTiger API (#5060)
* TXN_API_END_RETRY doesn't need to test WT_TXN_ERROR separately from ret being non-zero,
WT_TXN_ERROR will only be set if API_END found a non-zero ret.

* __session_commit_transaction() errors can't jump to the err: label and return, an error during
commit implies a rollback.

* There was a recent bug fix in the __session_commit_transaction() function to never skip
over the commit/rollback call regardless of the error. With that change, it's no longer
possible for WT_SESSION_TXN_STATE(session)->id to be anything other than WT_TXN_NONE.

* Check WT_TXN_ERROR before proceeding with any call: This isn't a correctness check,
if the WT_TXN_ERROR flag is set, we'll rollback the transaction no matter what the
application does. The reason for the change is to ensure we log an error so the
application knows it made a mistake, and second, to do it as soon as possible so
the application knows where to look. Assert in diagnostic builds that's the case.
    That fix requires a change to complex table cursor APPLY_CG macro so
it immediately fails if it sees any error other than WT_NOTFOUND, so rather than
repeatedly working through the list of table cursors, it quits as soon as there's
anything other than a not-found error.

Fix a problem where attempting an operation after a transaction has been prepared
can make it impossible to commit the prepared transaction. In short, attempting
an unsupported operation after a transaction is successfully prepared cannot set
the WT_TXN_ERROR flag, otherwise a subsequent attempt to commit the transaction
will instead roll it back. This requires we check if a transaction is prepared
(and return an error if it's not), before using the standard XXX_CONF macros,
because the corresponding API_END macro will will set the WT_TXN_ERROR flag if
the ret variable is non-zero.
    That fix requires a bunch of cursor function restructuring so functions
that return void can ignore the returned value from the prepared transaction
check function. In general, the pattern is to have a worker function that returns
an int and does all of the work, and a function that calls the worker function
and ignores its return. We need that pattern for other reasons in some places,
so this change just makes that a more common pattern to follow.
    That fix also requires passing in the operation "name" to the checking
function because we won't have the WT_SESSION handle set up before the call.
    That fix also requires minor changes to the static test suite where we were
attempting operations after a transactional call had already failed.

Move __wt_txn_context_prepare_check() and __wt_txn_context_check() from misc.i
to txn.i, the transaction subsystem has its own inlined function file.

* Alphabetize

* Rework the scope test: now we assert in DIAGNOSTIC mode if additional operations
are attempted inside a transaction after any operation fails, we have to limit
transactions to a single failed operation, and we have to rollback any transaction
where an operation fails.

Rework the main loop: clean up after every operation rather than trying to figure
out if the table is already in the right state.

* Check for WT_TXN_ERROR in API_SESSION_INIT, it's the only place to catch all of the
APIs. We have to permit WT_SESSION.rollback_transaction after an error occurs, but
it's the only one.

Check for fatal errors before initialization, it doesn't really matter but it's cleaner.

Fix a typo (which triggered a clang re-format).

* Rework api macros to set up standard names before calling __wt_txn_context_prepare_check()
so that naming works correctly in the standard error handlers.

* Now that the upper-level API code is checking WT_TXN_ERROR, the prepare function won't be called at
all if WT_TXN_ERROR is set. (This code doesn't look right anyway, it was likely copied-and-pasted
from the commit code, there's no reason for prepare to check if there were updates to the transaction.
Similarly, remove the assert that the transaction state has been reset, that only happens if the
prepare is successful, again, likely copied from commit.)

Add a comment to explain why commit is checking TXN.mod_count.

* Another related change: there's no reason for prepare to continue if the transaction has failed,
regardless of whether or not it's readonly.

* Generalize API_SESSION_PUSH/API_SESSION_POP to handle the WT_DATA_HANDLE push/pop as
well as the name and last-op information.

We can't return immediately out of SESSION_API_PREPARE_CHECK, we have to restore the
previous WT_SESSION information.

* There's no reason to print out WT_SESSION_IMPL.name explicitly, the error handling functions
will do that for you.