WT-5298 checkpoint to check validity window of history store leaf pages (#5033)
* WT-5298 Checkpoint to remove history store obsolete pages

Checkpoint has to check for the obsolete pages in history
store while checkpointing the history store by checking the
clean pages whether the stop time pair of the page makes it
all visible in the page or not? If all are visible on the page
means that the page is obsolete. Marke the obsolete page as
empty and their parent as dirty. During the parent page
processing, this page reference is removed in the parent thus
this page is no more referenced in the new checkpoint.

Whenever the old checkpoint that refers this page deleted
automatically these pages are deleted.

* Remove the start/stop time pair from WT_REF

This commit is somewhat revert of eaa0fe1f5b7a2ec9c3379b0e41ed73c162bee694
Instead of full revert, kept the structure declaration of WT_TIME_PAIR from
that commit.