WT-5321 Stats update according to durable history
As part of durable history project, the cache overflow
funcionatily is removed and new functionality of history
store is added. As part of this change, some of the existing
stats related to lookaside are invalid and those gets removed.
But these are stats are quite important to find out the performance
problems in the customer environments, because of that reason,
we will maintain some stats that are possible to provide with
history store useful for finding out performance issues.