WT-5377 variable-length column store insert locking in format test program can stall eviction (#5090)
* In variable-length column-store runs with sufficiently high insert percentages, the
global insert lock is stalling eviction because the thread holding the lock is helping
with eviction, and the threads pinning eviction are waiting on the lock. Move the list
of inserted records into a per-thread structure so there's no need to lock.

* Test after every operation if we have pending inserts so we avoid starvation should a
thread not insert soon after failing to update the total rows.