WT-5392 Adding history store page back to urgent queue if eviction fails
The obsolete history store pages that are added to evition is not able to evict
because of evicton threads try to evict the page before the checkpoint release
the hazard pointer on them. Try to add them back to the urgent queue until they
get evicted.

Added additional checks to avoid edge cases where the child page is deleted, but
not yet removed from internal page.

New statistic is added to find out how the eviction is performing with adding back
the obsolete page to the queue again.

Reorganize the tests to reduce the duplicate code