WT-5321 Stats name and description update according to durable history (#5093)
As part of the durable history project, the cache overflow
functionality is removed and new functionality of the history
store is added. As part of this change, some of the existing
stats related to lookaside are invalid and those are removed.
Some of these are stats are quite important to find out the performance
problems in customer environments. we are maintaining those stats
that are possible to provide with history store useful for finding out
performance issues.