WT-5379 missing varargs cleanup (#5095)
* The CURSOR_API_CALL macro can now both return or jump to the error label, so the stdarg
variable argument list support and cleanup needs to be moved to after those calls so we
get consistent behavior.

* The CURSOR_API_CALL macro can now both return or jump to the error label, so the stdarg
variable argument list support and cleanup needs to be moved to after those calls so we
get consistent behavior.

* The CURSOR_API_CALL() changes require multiple va_start/va_end pairs in a function, which breaks
the test for a va_start with a skipped va_end, the grep function does farthest match and I don't
see any easy way to fix it. Remove that test.