Add a global start time.
Add an timer-expired check function, and rework the main format loop to check if any configured
run-time has expired before doing any actions that are likely to take a long time to run. Assert
we accomplish at least one full set of format work, and if we haven't, fail the run.

Convert the main timer loop to use the library __wt_seconds() function instead of the more
complicated difftime() with time_t types.

Remove the format loop code intended for testing only a bulk-load of the data, that hasn't
been useful for quite a few years.

Add a timer-expired check to the bulk load function because it's the first place that stresses
the I/O subsystem, and we can run forever if the I/O system can't handle the load.