WT-5320 Update restore to do lookaside as well (#5107)
This change writes history versions to lookaside when doing update restore eviction. As a result, we combine the code path of update restore eviction and lookaside eviction and using the leave_dirty flag to indicate whether we need to restore the update chain if there are uncommitted changes.

We removed the WT_REC_UPDATE_RESTORE flag and used WT_REC_IN_MEMORY flag to avoid in-memory database also writing history versions to lookaside as for in-memory database writing to lookaside only results to more memory pressure.