WT-5392 Adding obsolete pages to eviction (#5151)
History store in-memory pages are not referenced
by any session in WT, evicting them in this checkpoint let them
to be cleaned in the next checkpoint. Store all the in-memory
obsolete pages in a list and add them to the urgent eviction queue at
the end of the checkpoint of history store.