WT-5219 Btree walk code read the lock WT_REF.addr field without locking (#5167)
* Add a locking version of __wt_ref_info() named __wt_ref_info_lock() that first locks the WT_REF
and then calls __wt_ref_info().

Change both versions of __wt_ref_info() to return a boolean if an address is known to reference
a leaf page, rather than returning a cell type.

Rework the compaction tests to check if a page needs to be re-written in two ways. First,
don't lock out checkpoints if a page hasn't been modified. If the page hasn't been modified,
we're checking the original addresses for the page and we only need to block checkpoints
when considering the page modify structure's addresses as set during reconciliation. Second,
use the new locking version of the __wt_ref_info() function to replace local locking of the
WT_REF structure, and give up the WT_REF lock immediately rather than hold it until the block
manager's compact_page_skip function returns.

Rework the page-walk __ref_is_leaf() function to use the locking version of the __wt_ref_info()
function, we don't hold a hazard pointer on the WT_REF, nor do we have it locked.

Move the __wt_page_addr_string() function into the page-verify code and rename it
__verify_addr_string(), that's the only place that calls it.