WT-5513 Don't consider TOMBSTONE updates for history store
Every history update must have a TOMBSTONE or the on-disk
record with a valid stop time pair. The TOMBSTONE update or
the stop time pair should be ignored for the history store
transaction read and later in their corresponding functions
they can be handled. Otherwise, none of the history store
entries are visible to the some of the transaction reads.