Use update flag to indicate the update is restored instead of connection wide flag
During rollback to stable to abort the newer on-disk values, we add an update from
history store. But during reconciliation, there are chances of the original value
gets appended to the update list. As it is a newer update and we don't need it.

we add a new flag in the update to indicate the update has been restored from
history store, so whenever during the append of original value, if any update
in the update chain has this flag, so we never append the on-disk value.