WT-5421 Replace an on-disk value with history store update (#5198)
Add a TOMBSTONE or STANDARD to update on-disk KV pairs. 
TOMBSTONE with WT_TS_NONE and WT_TXN_NONE to remove an on-disk key
and to remove an on-disk key remove operation, the STANDARD update
with the original timestamp and transaction id are added.

During rollback to stable to abort the newer on-disk values, we add an update from
history store. But during reconciliation, there are chances of the original value
gets appended to the update list. As it is a newer update and we don't need it.
To avoid it, a new flag in the update is added to indicate the update has been
restored from history store, so that on-disk original value never gets appended to
the update list.

When moving update from history store to data store, clear the transaction id
when recovery is in progress.