WT-5521 Cache stuck during format initial load, configured with library checkpoints (#5233)
* If reconciliation requires multiple blocks and checkpoint is running we'll eventually fail.
It's possible this is a big page that will take a lot of writes, avoid wasted work.

* Quit doing so much work in format's read-scan, it's not that useful any more and we're have already
verified the load.

* Configuring WiredTiger library checkpoints is done separately, rather than as part of the
original database open because format tests small caches and you can get into cache stuck
trouble during bulk load. Imagine a single thread doing lots of inserts and creating huge
leaf pages. Those pages can't be evicted when there are checkpoints running in the tree and
the cache gets stuck. That workload is unlikely enough the library can't handle it, and we
configure it away here.