WT-5612 History store cleanup for non-timestamped tables
There is no need of any history store values for non-timestamped
tables when there are no old running transactions in parallel as
the newer transactions can see the newer data that is visible in
the data store. So clean the history store for the non-timestamped
tables whenever rollback to stable is called.