WT-5583 Fix a dead lock in eviction server flow
In the eviction server code flow when there are no eviction workers,
the server itself try to evict the pages, during the page eviction
flow it try to take the lock which itself is already held leads to
a dead lock.

Fix is to release the lock before evicting the page from the queue
and take back the lock once it finish processing the queue.