WT-5690 Skip removing update from hs when its newer version is stable (#5336)
As part of rollback to stable process, it aborts the updates in the
order of update list, on page and reconciled page. As part of one page
value correction once the history updates are removed, don't remove
the visible updates again as part of processing reconciled page if the
newer version of the key is stable.