WT-5500 Implement new history store format (#5328)
* WT-5500 Initial work to support the new history store format

* Use the btree's write gen as the counter.

* Cleanup.

* Stop using write gen as the history store counter. Multiple records
with different transaction ids and the same timestamp will potentially
be written as part of the same btree write gen. For the time being,
I'm using a simple atomic counter. This isn't the final design but I'm
going to keep it this way while I debug the remaining issues and then
tackle this bit at the end.

* Fix a few more usages where we're searching history store.

* Remove unnecessary time pairs.

* Revert accidental break => goto.

* And the other instance.

* Create temporary variable to declutter atomic incr line.

* Update comment documenting the history store schema.

* Uncomment rollback to stable checks that were previously not working.

* Keep a per-btree counter instead of a global one for history store inserts.

* Fix compilation in Evergreen.

* Fix bad conflict resolution.

* Remove prepare state from the history store value. Prepared updates
never go to the history store. Once we support for evicting prepared
updates, they will always go on the page since they are the newest
version for a given key by definition.