WT-5631 Make sure to follow same connection config in the entire test (#5353)
Based on connection level logging enabled/disabled, the immediate durability
of the table is getting changed. Currently WT, doesn't save the configuration
in most of the scenarios, so make sure to use the same configuratio in the
entire test to let the behaviour of the tables same across the test.