WT-5680 On evicting an empty page, mark it DELETED not DISK (#5355)
When a remove becomes stable, stop writing keys even if some history is saved. Then when a page becomes empty, it won't have any content and will have a NULL address. These pages should be evicted into the DELETED state.

There was left over code for LIMBO pages (where a page could be empty in a checkpoint but have relevant history): this commit removes that code.