WT-5649 Refactor WT_REF locking, review all WT_REF.addr reads for locking issues (#5357)
* Remove the reconciliation locking around updating the WT_REF.addr fields, the consumers
of the WT_REF.addr information own ensuring they don't race with reconciliation.

Fix compact, when examining in-memory pages, to not read the modify structures or
WT_REF.addr field until it has blocked both eviction and checkpoint.
Fix compact, when examining on-disk addresses, to not read the WT_REF.addr until it
has locked the WT_REF structure.

Fix __sync_ref_obsolete_check(), when examining on-disk addresses, to not read the
WT_REF.addr until it has locked the WT_REF structure. Fix __sync_ref_obsolete_check(),
when examining in-memory pages, to not read the modify structures or the WT_REF.addr
field until it has acquired a hazard pointer to block eviction.

    Simplify __sync_dup_walk() and __sync_dup_hazard_pointer(), the latter doesn't
    need to return a WT_REF to its caller.

    Simplify __sync_ref_int_obsolete_cleanup() and __sync_ref_obsolete_check(), the
    latter doesn't need to increment the statistics on every  loop.

Replace __wt_ref_info_all() and __wt_ref_info() with __wt_ref_addr_copy() which
copies out the WT_REF.WT_ADDR structure address and timestamps into a new structure,
WT_ADDR_COPY. This isn't strictly a bug fix, but it makes it less likely that a caller
will release relevant locks before looking at the WT_REF.addr field. Additionally,
we no longer need to pass 6 arguments or check if they are NULL, it simplifies the code.

    Remove the __wt_ref_info_lock() function entirely, there was only a single caller,
    the compaction code.

Rework the btree debug page display code use the __debug_ref() function to display
WT_REF structures, and change __debug_ref() to display all of the WT_REF timestamps.

Rework __wt_addr_string() to check both the passed-in addr size as well as the reference,
just to be sure we don't attempt to crack an uninitialized WT_REF address.

Rework __verify_addr_string() to not treat the root as a special case, it still has the
usual WT_REF address information.

Add the usual __func__/__LINE__ macro for __wt_hazard_set to minimize how much clang-format
messes with comments.

* Add WT_REF_LOCK/WT_REF_UNLOCK macro pair, rework obvious places to use it.

* To look at an on-page cell, we need to look at the parent page's disk image, and that can be
dangerous. The problem is if the parent page splits, deepening the tree. As part of that
process, the WT_REF WT_ADDRs pointing into the parent's disk image are copied into off-page
WT_ADDRs and swapped into place. The content of the two WT_ADDRs are identical, and we don't
care which version we get as long as we don't mix-and-match the two.

Change the fast-delete code to use the common routine for copying the WT_REF.addr information,
and include the fix that was in the fast-delete code in the common routine.

* bt_delete.c:116:19: warning: The left operand of '!=' is a garbage value
if (addr.type != WT_ADDR_LEAF_NO)

* bt_sync.c:301:5: warning: Function call argument is an uninitialized value