WT-5696 Reconciliation use commit timestamp of cell as durable timestamp (#5373)
Currently, the value cell doesn't store the durable timestamp and when a page
is getting reconciled again with no newer updates on the page to be considered,
the aggregated durable timestamp is calculated as zero. This gets fixed when
the value cell has durable timestamp, till then consider the commit timestamp
instead of zero.