If we have no update to write, we have to discard any overflow value. Rework to do it all in one place by pointing the update to a fake tombstone record.
Remove the test for any update or insert when testing a specific key/value pairs
visibility, we don't care if there are associated updates or nearby inserts when
determining if we need to write any particular key/value pair.

Move the "dictionary" initialization out of the way, to the top of the loop.