WT-5684 overflow values must be discarded when there is no update for a key (#5374)
* If we have no update to write, we have to discard any overflow value. Rework to do it
all in one place by pointing the update to a fake tombstone record.

Remove the test for any update or insert when testing a specific key/value pairs
visibility, we don't care if there are associated updates or nearby inserts when
determining if we need to write any particular key/value pair.

Move the "dictionary" initialization out of the way, to the top of the loop.

* We use the tmpkey structure as part of removing the HS key, which is going to collide with
prefix compression if we don't reset the size to 0 since we never wrote the key. Move the
HS key clear code up into the tombstone handling code, there's no reason to wait until after
we've handled subsequent inserted keys before handling HS store removal.