WT-5780 Avoid inserting updates into history store with same start/stop timestamp
In case of out-of-order timestamp inserts from the application, it is possible
that newer update may have lesser timestamp than the current update. In such
scenarios don't insert them into the history store.

Make sure that history store stop timestamp is from the durable timestamp of a
newer update of the key insead of commit timestamp.