WT-5672 Support for start/stop aggregated durable timestamps
The rollback to stable operation depends on the aggregated
timestamps that are stored as part of the checkpoint and in the
internal pages of the tree. This change adjusts/corrects the
aggregation of both start/stop durable timestamps.

Along with aggregation support fixes, the rollback to stable
is modified to check for durable start/stop timestamp while
deciding the visibility of the value.