Add the check for historic page versions back in.
Minor restructuring to skip timestamp validity checks in the 4.2 branch when creating on-page
cells. It's no longer interesting to debug timestamp cell creation in the 4.2 branch, and
the change means we don't have to do any work to ensure reasonable durable timestamp values
get passed into the cell code in the 4.2 branch, we can just pass WT_TS_NONE and not worry
about it. (We will still continue to do full validity checking of any cells we read, which
might potentially be useful when debugging downgrades.)