WT-5923 Avoid performing history store cleanup for modified trees (#5451)
Scenarios when the trees are not checkpointed at once, it is possible
that durable timestamp of that tree can be zero, because of this reason
the trees are treated as non-timestamped tables and lead them to clean
their history store. To avoid this problem, ignore history store cleanup
when the trees are modified.

The side effect of this change is that it doesn't perform the history store
cleanup for the non-timestamped tables also that are modified, but anyway
these tables get cleaned during recovery rollback to stable phase.