WT-5931 Don't check the on-disk key/value when there is a valid update in update list
When the page was processed once for the rollback to stable and valid value is
restored from the history store and added to update list. Once again the page
is getting processed for rollback to stable, it must not process the on-disk
key/value again as it can lead to remove of the key.