WT-5931 Recovery rollback to stable is leading to remove a record that is not expected (#5460)
When the page was processed once for the rollback to stable and valid value is
restored from the history store and added to the update-list. Once again the page
is getting processed for rollback to stable, it must not process the on-disk
key/value again as it can lead to remove of the key.