WT-5932 RTS Aborting on-disk updates for in-memory database (#5469)
For in-memory databases, even after writing an update to disk,
it still retian all the older updates in the memory until they
are obsolete because there is no history store involved for
in-memory databaes. For rollback to stable to work on in-memory
databases, we need to clear the on-disk update whenever there
are no visible updates found in the update list according to
the stable timestamp.