WT-5948 Search shouldn't ignore globally visible tombstones (#5479)
As part of rollback to stable operations, any updates that are needs
to be removed, rollback to stable adds globally visible updates to
both data and history store. Search is not ignoring these updates
for the data store, but whereas these are getting ignored for history
store when session is requested to ignore the tombstones of the
history store, but globally visible tombstone updates shouldn't be
ignored for hs irrespective of the session flags.