Use parent session no logging flags to rts internal session
Using the parent session no logging flags to the internal session
that is created by the rollback to stable operation to not let the
log records to be created as the parent session is not insisted to
log records. Recovery flow rollback to stable is the one of the flow
where log records must not be generated.