Changes to support reopening existing databases.
General renaming of underlying objects for consistency; remove Win32 support as part of this
change, we don't run format on Win32 at the moment and some of the existing support was wrong.

Use a timestamp to guarantee backup ID uniqueness, otherwise we fail when we reopen the
database.

Write out a list of the randomly chosen key lengths so we can read it back in on reopen.