WT-5766 Separate out internal and shared transaction data (#5485)
Moved transaction state that is read from non-owning sessions to a separate WT_TXN_SHARED structure and use this for the read and durable queues. This means that after being placed in the read and durable queues, this shared transaction state is not written to.

The main functional change here is that we've added a pinned read timestamp which is used to calculate how far we can move oldest. We were previously hitting a problem when reading history store modifies since we need to shift our read timestamp forward to read earlier modifies or the base update itself. If the oldest timestamp is updated during that process, it's possible to obsolete data that we aren't allowed to yet. This issue is solved by maintaining an immutable pinned read timestamp in the shared transaction state and only modifying the read timestamp of the local transaction state.

Co-authored-by: Michael Cahill <michael.cahill@mongodb.com>
Co-authored-by: Keith Bostic <keith.bostic@mongodb.com>
(cherry picked from commit 2f13fa75ea16abb06b3abbdcf8481364bcce5514)