Simplify/clarify __rec_append_orig_value(), mostly fixing comments, no substantive changes.
Do cheap checks first when reviewing the update list.

Allocate append before tombstone, it's simpler and avoids extra branching.

No need to check unpack == NULL in the code, we assert in the header it's not NULL. Remove
comments that discuss "vpack" being optional, we're guaranteed a renamed vpack in this code.

Remove [-Wconditional-uninitialized] comments: any warnings were correct, the variables could
be used uninitialized if not initialized explicitly