WT-3726 Add documentation specifying build prerequisites
Added INSTALL file (and script to create it) pointing users
to the "Building and installing WiredTiger on POSIX" and
"Building and installing WiredTiger on Windows" pages.  These
seem to have the appropriate information about prerequisites, as
well as instructions for building and installing, which are often
found in open source INSTALL files.

Updated the Building on POSIX page to include SWIG as a
prereq for building Python and Java support.