WT-5833 overflow key/value item caching is not working (#5523)
* Make a copy of the on-page value if the value is an overflow item and anything other than
the on-page cell is being written. This is because the value's backing overflow blocks
aren't part of the page, and they are physically removed by checkpoint writing this page,
that is, the checkpoint doesn't include the overflow blocks so they're removed and future
readers of this page won't be able to find them.

* Fix a typo in a comment.

* The per-page overflow cache is not sufficient for eviction now that pages with deleted overflow
records can be evicted, even if there are snapshot readers that require the previous values.
Such deleted overflow records must be copied into the HS file. Since the HS file also acts as
a cache for checkpoint deleted overflow records, remove the per-page overflow cache.

* Update an out of date comment

Co-authored-by: Chenhao Qu <chenhao.qu@mongodb.com>