WT-5749 Only pass a non-NULL WT_ITEM value when allocating a WT_UPDATE with associated data (#5529)
* Rename __wt_update_alloc() to __wt_upd_alloc() so naming is consistent with
__wt_upd_alloc_tombstone().

Change __wt_upd_alloc() argument order so all of the returned side-effects are the last arguments
passed to the function.  Move __wt_upd_alloc() into txn.i, we call it a lot, and that's where the
__wt_upd_alloc_tombstone() function lives.

Convert all code that allocates tombstones to call __wt_upd_alloc_tombstone().

Add a size argument to __wt_upd_alloc_tombstone(), replace code that passed in size arguments, but
never used the return value, to pass in a NULL.

Assert reserved and tombstone updates are never allocated with a non-NULL value argument.

Restructure Btree cursor code to not pass a value when allocating reserved or tombstone updates.

Minor cleanup in __wt_col_modify() to simplify a test.

* Fix up the initial assert in col/row-modify to include WT_UPDATE_RESERVE in the case of the
value or update being supplied.