WT-5911 Remove unsafe calls to verify HS from session->verify. (#5499)
* WT-5911 Remove unsafe calls to verify HS from session->verify.

* Fix free of uri_data. Remove unused function.

* Shorten function name. Fix initialization.

* Move verification to end of work loop but before done label.

* Don't verify the history store during metadata salvage yet.

* Refactor calls to verify history store.

* Call history store verify while processing the last checkpoint.

* Add 'verify_metadata' config string.

* Fix comments.

* Reset cursor to release resources.

* Release each page searched for once we no longer need it, we've acquired a hazard
pointer during the search.

Don't both calling both __wt_btcur_reset() and __wt_btcur_close(), just don't set
the low-level flag so the cursor is reset. This is probably not required in this
path, but it doesn't hurt and could potentially avoid bugs in error paths.

Fix a comment typo.

* Don't open a cursor unless we need one.
Don't allocate a scratch buffer, we don't need one.

* Don't copy the buffer's contents, just swap the pointers.

* Don't verify the file against HS while eviction is locked out, we're potentially reading a lot
of pages into memory, we need to let eviction run.

We can use the fact the checkpoints are time-ordered to avoid getting and saving a copy of the
final checkpoint's name. (I also don't think this is safe in the context of named checkpoints,
__wt_meta_checkpoint_last_name() only cares about unnamed checkpoints and I think we could get
into trouble here.

* I left a use of the cursor after it had been reset (it's simpler this way because of error
handling, so get to the data-handle via the session instead of the cursor).

* Remove debugging statements.

* Return WT_NOTFOUND.
Add -m option to wt to verify metadata (not working yet).

* Whitespace and strings

* wt_row_search does not return WT_NOTFOUND.

* Use internal session for whole HS verification.

* Open internal session around verifying metadata.

* Skip history store verify of history store and metadata.

* Set false to not open a metadata cursor on the internal session.

* Move where history store is verified.

* Check for WT_NOTFOUND.

* Whitespace

* Return EINVAL instead of WT_NOTFOUND on actual failure. WT_NOTFOUND is
used for detecting the end of the cursor/table walk.

* Return WT_ERROR instead of EINVAL.

* Address review comments.

* Move verify of history store to after eviction is started.

* Simplify error handling.

Co-authored-by: Keith Bostic <keith.bostic@mongodb.com>